+++ src/CSRakowski.Parallel/ParallelAsync.Unordered.cs
@@ -22,7 +22,7 @@ namespace CSRakowski.Parallel
             using (var enumerator = collection.GetEnumerator())
             {
                 var hasNext = true;
-                long batchId = 0;
+                int batchId = 0;
                 var taskList = ListHelpers.GetList<Task<TResult>>(batchSize);
 
                 while (!cancellationToken.IsCancellationRequested)
@@ -77,7 +77,7 @@ namespace CSRakowski.Parallel
             using (var enumerator = collection.GetEnumerator())
             {
                 var hasNext = true;
-                long batchId = 0;
+                int batchId = 0;
                 var taskList = ListHelpers.GetList<Task>(batchSize);
 
                 while (!cancellationToken.IsCancellationRequested)
